server/app/utils/auth_middleware.py
from functools import wraps
from flask import request, jsonify
import jwt
from app.config import Config

def verify_token(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None

        # Check if token is in headers
        auth_header = request.headers.get('Authorization')
        if auth_header:
            try:
                token = auth_header.split(" ")[1]  # Bearer <token>
            except IndexError:
                return jsonify({'error': 'Invalid token format'}), 401

        if not token:
            return jsonify({'error': 'Token is missing'}), 401

        try:
            # Decode token
            data = jwt.decode(token, Config.JWT_SECRET_KEY, algorithms=["HS256"])
            request.current_user = {
                'user_id': data['user_id'],
                'email': data['email'],
                'roles': data.get('roles', ['student'])
            }
            return f(*args, **kwargs)
        except jwt.ExpiredSignatureError:
            return jsonify({'error': 'Token has expired'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'error': 'Invalid token'}), 401
        except Exception as e:
            return jsonify({'error': str(e)}), 401

    return decorated

def check_role(allowed_roles):
    def decorator(f):
        @wraps(f)
        def decorated(*args, **kwargs):
            if not request.current_user:
                return jsonify({'error': 'Authentication required'}), 401
                
            user_roles = set(request.current_user.get('roles', ['student']))
            required_roles = set(allowed_roles)
            
            if not user_roles.intersection(required_roles):
                return jsonify({'error': 'Insufficient permissions'}), 403
                
            return f(*args, **kwargs)
        return decorated
    return decorator 