server/requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
flask-cors==3.0.10
python-dotenv==0.19.0
PyJWT==2.1.0
werkzeug==2.0.1
python-dateutil==2.8.2
scikit-learn==0.24.2
numpy==1.21.2 